refactor: Move files Host into src
