+++ src/Host/appsettings.Development.json
@@ -1,12 +1,6 @@
 {
   "ConnectionStrings": {
-    "Marten": "Host=authdev-db;Port=5432;Database=AuthDev;Username=postgres;Password=postgres"
-  },
-  "AuthKit": {
-    "MaxDeveloperTokens": 3
-  },
-  "Encryption": {
-    "AES_MASTER_KEY": "Tww4kOE+310tZvqXvw1tg0779qccuy8t4vrYthcPP7c="
+    "DefaultConnection": "Host=auth-db;Port=5432;Database=Auth;Username=postgres;Password=postgres"
   },
   "Server": {
     "Host": "http://0.0.0.0:8080",
@@ -29,9 +23,7 @@
       ".ValueObject"
     ],
     "AllowedLayers": [
-      "Application",
-      "Domain",
-      "Infrastructure",
+      "Core",
       "Host"
     ],
     "SkipInterfaces": true,
@@ -40,14 +32,8 @@
   },
   "Logging": {
     "LogLevel": {
-      "Default": "Information",
+      "Default": "Debug",
       "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "AllowedHosts": "*",
-  "Kestrel": {
-    "EndpointDefaults": {
-      "Protocols": "Http2"
     }
   }
 }
