+++ src/Host/appsettings.json
@@ -1,6 +1,12 @@
 {
   "ConnectionStrings": {
-    "DefaultConnection": "Host=auth-db;Port=5432;Database=Auth;Username=postgres;Password=postgres"
+    "Marten": "Host=authdev-db;Port=5432;Database=AuthDev;Username=postgres;Password=postgres"
+  },
+  "AuthKit": {
+    "MaxDeveloperTokens": 3
+  },
+  "Encryption": {
+    "AES_MASTER_KEY": "Tww4kOE+310tZvqXvw1tg0779qccuy8t4vrYthcPP7c="
   },
   "Server": {
     "Host": "http://0.0.0.0:8080",
@@ -23,9 +29,7 @@
       ".ValueObject"
     ],
     "AllowedLayers": [
-      "Application",
-      "Domain",
-      "Infrastructure",
+      "Core",
       "Host"
     ],
     "SkipInterfaces": true,
@@ -34,8 +38,14 @@
   },
   "Logging": {
     "LogLevel": {
-      "Default": "Debug",
+      "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "AllowedHosts": "*",
+  "Kestrel": {
+    "EndpointDefaults": {
+      "Protocols": "Http2"
     }
   }
 }
